4.2.2: use long not int (overflow)

+++ main/phylonet/coalescent/CalculateWeightTask.java
@@ -1,5 +1,5 @@
 package phylonet.coalescent;
 
 interface CalculateWeightTask<T> {
-	Integer calculateWeight();
+	Long calculateWeight();
 }
